+++ src/views/PluginEdit.vue
@@ -3,13 +3,27 @@
     <portal to="app-bar-extension">
       <div style="display: flex; flex-direction: row; width: 100%">
         <v-tabs v-model="tab" align-with-title>
-          <v-tab key="interact" href="#interact">
+          <v-tab
+            key="interact"
+            href="#interact"
+            :disabled="
+              !plugin.loaded || !plugin.enabled || !plugin.execution_enabled
+            "
+          >
             Interact
             <v-icon x-small class="ml-1"> fa-terminal </v-icon>
           </v-tab>
-          <v-tab key="tasks" href="#tasks">
+          <v-tab key="tasks" href="#tasks" :disabled="!plugin.loaded">
             Tasks
             <v-icon x-small class="ml-1"> fa-sticky-note </v-icon>
+          </v-tab>
+          <v-tab key="details" href="#details">
+            Details
+            <v-icon x-small class="ml-1"> fa-info-circle </v-icon>
+          </v-tab>
+          <v-tab key="settings" href="#settings" :disabled="!plugin.loaded">
+            Settings
+            <v-icon x-small class="ml-1"> fa-cog </v-icon>
           </v-tab>
         </v-tabs>
       </div>
@@ -22,13 +36,22 @@
       :submit-loading="loading"
       @submit="submit"
     >
-      <template slot="extra-stuff">
+      <template #extra-stuff>
         <tooltip-button-toggle
+          v-if="plugin.loaded"
           v-model="isRefreshTasks"
           icon="fa-redo"
           :button-text="isRefreshTasks ? 'On' : 'Off'"
           text="Auto-refresh Tasks"
         />
+        <v-switch
+          v-model="plugin.enabled"
+          :disabled="!plugin.loaded"
+          color="green"
+          label="Enabled"
+          class="mr-2 mt-3"
+          @change="toggleEnabled"
+        ></v-switch>
       </template>
     </edit-page-top>
     <error-state-alert
@@ -45,15 +68,19 @@
       >
         <h4 class="pl-4 pb-4">Execute Plugin</h4>
         <v-card style="padding: 10px">
-          <info-viewer class="info-viewer" :info="pluginInfo" />
           <technique-chips :techniques="plugin.TechniqueChips" />
           <general-form
             v-if="reset"
             ref="generalform"
             v-model="form"
-            :options="plugin.options"
+            :options="pluginOptions"
           />
-          <v-btn :loading="loading" color="primary" @click="submit">
+          <v-btn
+            :disabled="!plugin.enabled"
+            :loading="loading"
+            color="primary"
+            @click="submit"
+          >
             Submit
           </v-btn>
         </v-card>
@@ -68,38 +95,117 @@
           <plugin-tasks-list :plugin="plugin" :refresh-tasks="isRefreshTasks" />
         </v-card>
       </v-tab-item>
+      <v-tab-item
+        v-if="initialLoad"
+        key="details"
+        :value="'details'"
+        :transition="false"
+        :reverse-transition="false"
+      >
+        <v-card flat>
+          <v-card-title>Details</v-card-title>
+          <v-card-text>
+            <v-row>
+              <v-col cols="12">
+                <span v-if="!plugin.loaded">
+                  <v-alert prominent type="warning" outlined>
+                    <v-row align="center">
+                      <v-col class="grow">
+                        <vue-markdown
+                          :source="pluginPythonDeps"
+                          :options="{ html: true }"
+                        />
+                      </v-col>
+                    </v-row>
+                  </v-alert>
+                </span>
+                <v-spacer />
+                Authors:
+                <author-chips :authors="plugin.authors" />
+                <v-spacer />
+                <v-card
+                  v-if="plugin.readme"
+                  class="pa-4"
+                  outlined
+                  elevation="2"
+                >
+                  <vue-markdown :source="plugin.readme" />
+                </v-card>
+              </v-col>
+            </v-row>
+          </v-card-text>
+        </v-card>
+      </v-tab-item>
+      <v-tab-item
+        key="settings"
+        :value="'settings'"
+        :transition="false"
+        :reverse-transition="false"
+      >
+        <v-card flat>
+          <v-card-title>Settings</v-card-title>
+          <v-card-text>
+            <v-row>
+              <v-col cols="12">
+                <span v-if="Object.keys(plugin.settings_options || {}) == 0"
+                  >No settings available</span
+                >
+                <template v-else>
+                  <general-form
+                    v-if="reset"
+                    ref="settingsform"
+                    v-model="settingsForm"
+                    :options="plugin.settings_options"
+                  />
+                  <v-btn
+                    :loading="loading"
+                    color="primary"
+                    @click="submitSettings"
+                  >
+                    Submit
+                  </v-btn>
+                </template>
+              </v-col>
+            </v-row>
+          </v-card-text>
+        </v-card>
+      </v-tab-item>
     </v-tabs-items>
   </div>
 </template>
 
 <script>
 import GeneralForm from "@/components/GeneralForm.vue";
-import InfoViewer from "@/components/InfoViewer.vue";
 import * as pluginApi from "@/api/plugin-api";
 import TechniqueChips from "@/components/TechniqueChips.vue";
 import EditPageTop from "@/components/EditPageTop.vue";
 import PluginTasksList from "@/components/plugins/PluginTasksList.vue";
 import TooltipButtonToggle from "@/components/TooltipButtonToggle.vue";
 import ErrorStateAlert from "@/components/ErrorStateAlert.vue";
+import VueMarkdown from "vue-markdown-render";
+import AuthorChips from "@/components/AuthorChips.vue";
 
 export default {
   name: "PluginEdit",
   components: {
-    InfoViewer,
+    AuthorChips,
     GeneralForm,
     TechniqueChips,
     EditPageTop,
     PluginTasksList,
     TooltipButtonToggle,
     ErrorStateAlert,
+    VueMarkdown,
   },
   data() {
     return {
       reset: true,
       loading: false,
       isRefreshTasks: false,
+      initialLoad: false,
       form: {},
-      plugin: { options: {} },
+      settingsForm: {},
+      plugin: { execution_options: {} },
       errorState: false,
     };
   },
@@ -133,10 +239,10 @@ export default {
       };
     },
     pluginOptions() {
-      const { options } = this.plugin;
-      if (!options) return {};
+      const { execution_options } = this.plugin;
+      if (!execution_options) return {};
 
-      return options;
+      return execution_options;
     },
     id() {
       return this.$route.params.id;
@@ -146,8 +252,30 @@ export default {
         this.$router.replace({ query: { ...this.$route.query, tab } });
       },
       get() {
-        return this.$route.query.tab || "interact";
+        if (this.$route.query.tab) {
+          return this.$route.query.tab;
+        }
+        if (!this.plugin?.execution_enabled) {
+          return "details";
+        }
+
+        return "interact";
       },
+    },
+    pluginPythonDeps() {
+      if (!this.plugin.python_deps.length > 0) {
+        return `
+Plugin wasn't loaded. Check the server logs.
+`;
+      }
+
+      return `
+This plugin requires additional Python dependencies.
+Please install and restart the server.
+\`\`\`sh
+poetry add ${this.plugin.python_deps.join(" ")}
+\`\`\`
+      `;
     },
   },
   mounted() {
@@ -163,7 +291,7 @@ export default {
 
       try {
         const response = await pluginApi.executePlugin(
-          this.plugin.name,
+          this.plugin.id,
           this.form,
         );
         this.$snack.success(`${response.detail}`);
@@ -172,6 +300,40 @@ export default {
       }
 
       this.loading = false;
+    },
+    async submitSettings() {
+      if (this.loading || !this.$refs.settingsform.$refs.form.validate()) {
+        return;
+      }
+
+      this.loading = true;
+
+      try {
+        const response = await pluginApi.updatePlugin({
+          ...this.plugin,
+          options: this.settingsForm,
+        });
+        this.plugin = response;
+        this.$snack.success("Settings updated");
+      } catch (err) {
+        this.$snack.error(`Error: ${err}`);
+      }
+
+      this.loading = false;
+    },
+    async toggleEnabled(val) {
+      this.plugin.enabled = val;
+
+      try {
+        const response = await pluginApi.updatePlugin({
+          ...this.plugin,
+          options: this.form,
+        });
+        this.plugin = response;
+      } catch (err) {
+        this.plugin.enabled = !val;
+        this.$snack.error(`Error: ${err}`);
+      }
     },
     getPlugin(id) {
       pluginApi
@@ -186,6 +348,9 @@ export default {
         })
         .catch(() => {
           this.errorState = true;
+        })
+        .finally(() => {
+          this.initialLoad = true;
         });
     },
   },